+ phase correction algorithm + code refactoring
